read emails: refactor client react code
